pyproject.toml
[tool.poetry]
name = "A2A_showcase"
version = "0.1.0"
description = "A2A showcase"
authors = ["Olusanmi Hundogan <o.hundogan@ohcm.nl>"]
license = "OHCM B.V. License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.4"
python-dotenv = "^1.0.0"
cloudpickle = "^3.0.0"
rich = "^14.0.0"
pyyaml = "^6.0.2"
pydantic-settings = "^2.10.1"
colorlog = "^6.9.0"
langchain = "^0.3.27"


[tool.poetry.group.agent.dependencies]
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"
streamlit = "^1.45.1"
langgraph = "<0.6" # TODO: Need to update to langgraph 0.6 or higher
langgraph-supervisor = "^0.0.27"
pydantic-settings = "^2.10.1"
a2a-sdk = "^0.2.8"
python-a2a = "^0.5.9"
langchain-mcp-adapters = "^0.1.9"

[tool.poetry.group.mcp.dependencies]
fastmcp = "^2.11.1"
duckduckgo-search = "^8.0.3"
bs4 = "^0.0.2"
serpapi = "^0.1.5"
langchain = "^0.3.27"
langchain-community = "^0.3.27"
aiosmtplib = "^3.0.1"
yfinance = "^0.2.65"

[tool.poetry.group.playwright.dependencies]
playwright = "^1.40.0"
pytest-playwright = "^0.7.0"
crawl4ai = "^0.6.3"

[tool.poetry.group.kafka.dependencies]
kafka = "^1.3.5"
kafka-python = "^2.2.14"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
yapf = "^0.40.0"
pylint = "^2.17.4"
ipykernel = "^6.29.5"



[tool.yapf]
based_on_style = "pep8"
column_limit = 150

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
